array type now stores multidimensional arrays
The array type also now has a method to permute its dimensions, this
should be useful for shuffling memory before copying it to the gpu.

Also removed operator& from the array type. Since that operator
overwrites the ability to take the address of an array it could have
caused issues later.