re-structured the renderer to sync cpu and gpu
Before the renderer was structured more like a game renderer that
will do a lot of cpu side work, so you want the cpu to continue
working while the gpu is rendering. This causes data race problems
though when you need to mess with buffers the gpu is still working
with, it seems this is usually solved by allocating multiple versions
of the asset on the gpu, one for each frame in flight. That's pretty
bad in this renderer though since the gpu will do almost all the work
and the buffers will be large so using two or more times the memory
to prevent data races isn't really possible. So the cpu and gpu should
just work serially which simplifies buffer management significantly
and uses less memory with minimal to no speed compromise.